Improved the UI (still text based), Added Sticky Key Support and amended SET SystemParametersInfo() calls to write the data to the user profile. An issue has been encountered where the program has slight issues with resolutions above 1080p
The UI has improved to have moore options and notices hopefully making the program slightly easier to use. Implemented the use of Sticky Keys for users who are used to toggling the ctrl/shift modifiers for mousekeys as opposed to holding them manually, however these modes still require users to toggle the correct keys themselves as the API doesn't appear to support it, awaiting a Microsoft developer ticket to find out. SET SystemParametersInfo() calls now write the amended data to the user profile allowing the indicators for MouseKeys and StickyKeys to actually show in the system tray. An issue has been encountered when the screen resolution is above 1080p, it would appear the mousekey movement amount doesn't scale for screens, this will be looked into.